resolving code comments to move helper logic to ```prometheus_helper.py```
